feat: Ensure correct order when toggling isGroup
